Major reimplementation to fix use-after-free
Previous implementation used multiple objects which all used the same
underlying xenstat_node structure. That strusture was tied with the Node
object. When that Node object went out of scope, it was possible that
the underlying xenstat_node could be freed by GC, causing use-after-free
in the other objects still referencing the xenstat_node structure.